moving playwright to plugin
