+++ packages/plugin-playwright/src/index.ts
@@ -1,3 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
+export * from "./types.js";
 export * from "./playwright.js";
+export * from "./browse.js";
